feat: exportImage 유틸함수로 분리 (#27)
